+++ packages/akutil-common/CHANGELOG.md
@@ -1,11 +1,16 @@
 <a name="0.1.0"></a>
-# 0.1.0 (2016-07-25)
+# 0.1.0 (2016-08-01)
 
 
 ### Bug Fixes
 
+* endsWith is not available in IE ([d8b1f01](https://bitbucket.org/atlassian/atlaskit/commits/d8b1f01))
+* FAB-370 add attributes to prop types in pf-mention-picker ([b1929da](https://bitbucket.org/atlassian/atlaskit/commits/b1929da))
+* Fix bad export for MentionResource ([1502f73](https://bitbucket.org/atlassian/atlaskit/commits/1502f73))
+* Remove broken export from ak-inline-dialog, and prevent usage upstream. ([e5cc55f](https://bitbucket.org/atlassian/atlaskit/commits/e5cc55f))
 * **skate:** SkateJS update to beta.10 ([620034a](https://bitbucket.org/atlassian/atlaskit/commits/620034a))
 * remove dependency to popper ([5e19bbe](https://bitbucket.org/atlassian/atlaskit/commits/5e19bbe))
+* target for alignment fix ([20d6cc2](https://bitbucket.org/atlassian/atlaskit/commits/20d6cc2))